make fp6_llm.py
