+++ torchao/dtypes/__init__.py
@@ -1,7 +1,7 @@
 from .nf4tensor import NF4Tensor, to_nf4
 from .uint4 import UInt4Tensor
 from .aqt import AffineQuantizedTensor, to_aq
-from .float6_e3m2 import to_float6_e3m2, from_float6_e3m2, to_tc_float6_e3m2, from_tc_float6_e3m2
+from .float6_e3m2 import to_float6_e3m2, from_float6_e3m2
 
 __all__ = [
     "NF4Tensor",
@@ -11,6 +11,4 @@ __all__ = [
     "to_aq",
     "to_float6_e3m2",
     "from_float6_e3m2",
-    "to_tc_float6_e3m2",
-    "from_tc_float6_e3m2",
 ]
